refactor(pollux): rework internal components
