+++ crates/pollux/src/main.rs
@@ -7,8 +7,6 @@ mod ioc;
 mod pollux;
 
 use crate::infra::cli;
-use crate::pollux::PolluxTask;
-use PolluxTask::*;
 use tikv_jemallocator::Jemalloc;
 
 #[global_allocator]
@@ -20,15 +18,7 @@ async fn main() -> anyhow::Result<()> {
     let pollux = ioc::create_pollux();
 
     let task = cli::parsing::parse_arguments()?;
-
-    match task {
-        EvaluateRustProject(project_root) => pollux.evaluate_local_project(project_root.as_path()).await?,
-        EvaluateRustCrate(cargo_package) => pollux.evaluate_crate_package(&cargo_package).await?,
-        CleanupEverything => pollux.cleanup_everything(),
-        CleanupPackages => pollux.cleanup_cached_packages(),
-        CleanupEvaluations => pollux.cleanup_cached_evaluations(),
-        CheckRustCrate(cargo_package) => pollux.check_crate_package(&cargo_package).await?,
-    }
+    pollux.execute(task).await?;
 
     Ok(())
 }
